Move jsonl_path fixture to conftest.py
