+++ tests/conftest.py
@@ -1,3 +1,4 @@
+import json
 import lzma
 import textwrap
 
@@ -81,3 +82,20 @@ def xml_file(tmp_path_factory):
     with open(filename, "w") as f:
         f.write(data)
     return filename
+
+
+DATA = [
+    {"col_1": "0", "col_2": 0, "col_3": 0.0},
+    {"col_1": "1", "col_2": 1, "col_3": 1.0},
+    {"col_1": "2", "col_2": 2, "col_3": 2.0},
+    {"col_1": "3", "col_2": 3, "col_3": 3.0},
+]
+
+
+@pytest.fixture(scope="session")
+def jsonl_path(tmp_path_factory):
+    path = str(tmp_path_factory.mktemp("data") / "dataset.jsonl")
+    with open(path, "w") as f:
+        for item in DATA:
+            f.write(json.dumps(item))
+    return path
