+++ tests/io/test_json.py
@@ -1,26 +1,7 @@
-import json
-
 import pytest
 
 from datasets import Dataset, DatasetDict, Features, NamedSplit, Value
 from datasets.io.json import JsonDatasetReader
-
-
-DATA = [
-    {"col_1": "0", "col_2": 0, "col_3": 0.0},
-    {"col_1": "1", "col_2": 1, "col_3": 1.0},
-    {"col_1": "2", "col_2": 2, "col_3": 2.0},
-    {"col_1": "3", "col_2": 3, "col_3": 3.0},
-]
-
-
-@pytest.fixture(scope="session")
-def jsonl_path(tmp_path_factory):
-    path = str(tmp_path_factory.mktemp("data") / "dataset.jsonl")
-    with open(path, "w") as f:
-        for item in DATA:
-            f.write(json.dumps(item))
-    return path
 
 
 @pytest.mark.parametrize(
@@ -35,7 +16,7 @@ def jsonl_path(tmp_path_factory):
 )
 @pytest.mark.parametrize("split", [None, NamedSplit("train"), "train", "test"])
 @pytest.mark.parametrize("path_type", [str, list])
-def test_json_dataset_reader(path_type, split, jsonl_path, features, tmp_path):
+def test_json_dataset_reader(path_type, split, features, jsonl_path, tmp_path):
     if issubclass(path_type, str):
         path = jsonl_path
     elif issubclass(path_type, list):
@@ -48,14 +29,14 @@ def test_json_dataset_reader(path_type, split, jsonl_path, features, tmp_path):
     expected_features = features.copy() if features else default_expected_features
     features = Features({feature: Value(dtype) for feature, dtype in features.items()}) if features else None
 
-    ds = JsonDatasetReader(path, split=split, features=features, cache_dir=cache_dir).read()
-    assert isinstance(ds, Dataset)
-    assert ds.num_rows == 4
-    assert ds.num_columns == 3
-    assert ds.column_names == ["col_1", "col_2", "col_3"]
-    assert ds.split == expected_split
+    dataset = JsonDatasetReader(path, split=split, features=features, cache_dir=cache_dir).read()
+    assert isinstance(dataset, Dataset)
+    assert dataset.num_rows == 4
+    assert dataset.num_columns == 3
+    assert dataset.column_names == ["col_1", "col_2", "col_3"]
+    assert dataset.split == expected_split
     for feature, expected_dtype in expected_features.items():
-        assert ds.features[feature].dtype == expected_dtype
+        assert dataset.features[feature].dtype == expected_dtype
 
 
 @pytest.mark.parametrize(
@@ -69,7 +50,7 @@ def test_json_dataset_reader(path_type, split, jsonl_path, features, tmp_path):
     ],
 )
 @pytest.mark.parametrize("split", [None, "train", "test"])
-def test_json_datasetdict_reader(split, jsonl_path, features, tmp_path):
+def test_json_datasetdict_reader(split, features, jsonl_path, tmp_path):
     if split:
         path = {split: jsonl_path}
     else:
@@ -81,12 +62,12 @@ def test_json_datasetdict_reader(split, jsonl_path, features, tmp_path):
     expected_features = features.copy() if features else default_expected_features
     features = Features({feature: Value(dtype) for feature, dtype in features.items()}) if features else None
 
-    ds = JsonDatasetReader(path, features=features, cache_dir=cache_dir).read()
-    assert isinstance(ds, DatasetDict)
-    ds = ds[split]
-    assert ds.num_rows == 4
-    assert ds.num_columns == 3
-    assert ds.column_names == ["col_1", "col_2", "col_3"]
-    assert ds.split == split
+    dataset = JsonDatasetReader(path, features=features, cache_dir=cache_dir).read()
+    assert isinstance(dataset, DatasetDict)
+    dataset = dataset[split]
+    assert dataset.num_rows == 4
+    assert dataset.num_columns == 3
+    assert dataset.column_names == ["col_1", "col_2", "col_3"]
+    assert dataset.split == split
     for feature, expected_dtype in expected_features.items():
-        assert ds.features[feature].dtype == expected_dtype
+        assert dataset.features[feature].dtype == expected_dtype
